+++ src/tools-TEST/data_in/SplicingToBed_init.sql
@@ -2,10 +2,10 @@
 INSERT INTO `gene` (`id`, `hgnc_id`, `symbol`, `name`, `type`, `ensembl_id`) VALUES
 (22713,12692,'VIM','vimentin','protein-coding gene',NULL);
 
-INSERT INTO `gene_transcript` (`id`, `gene_id`, `name`, `source`, `chromosome`, `start_coding`, `end_coding`, `strand`) VALUES 
-(85651,22713,'ENST00000224237','ensembl','10',17271422,17279270,'+'),
-(85652,22713,'CCDS7120.1','ccds','10',17271422,17279270,'+'),
-(85653,22713,'ENST00000544301','ensembl','10',17271422,17279270,'+');
+INSERT INTO `gene_transcript` (`id`, `gene_id`, `name`, `version`, `source`, `chromosome`, `start_coding`, `end_coding`, `strand`) VALUES
+(85651,22713,'ENST00000224237','1','ensembl','10',17271422,17279270,'+'),
+(85652,22713,'CCDS7120','1','ccds','10',17271422,17279270,'+'),
+(85653,22713,'ENST00000544301','1','ensembl','10',17271422,17279270,'+');
 
 INSERT INTO `gene_exon` (`transcript_id`, `start`, `end`) VALUES
 (85651,17271277,17271984),